etc : build setting

+++ src/main/java/com/intouch/aligooligo/Routine/RoutineRepository.java
@@ -2,5 +2,5 @@ package com.intouch.aligooligo.Routine;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoutineRepository extends JpaRepository<Routine, Long> {
+public interface RoutineRepository extends JpaRepository<Routine, Integer> {
 }
