+++ src/main/java/com/intouch/aligooligo/Subgoal/SubgoalRepository.java
@@ -2,5 +2,5 @@ package com.intouch.aligooligo.Subgoal;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubgoalRepository extends JpaRepository<Subgoal, Long> {
+public interface SubgoalRepository extends JpaRepository<Subgoal, Integer> {
 }
